user-service-main/pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "User Service for Distributed Notification System"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]",
    "uvicorn[standard]",
    "pydantic",
    "pydantic-settings",
    "sqlalchemy",
    "psycopg2-binary",
    "alembic",
    "redis",
    "bcrypt",
    "python-jose[cryptography]",
    "python-multipart",
    "prometheus-client",
    "prometheus-fastapi-instrumentator",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
    "ruff",
    "fakeredis",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]